Add fixes from https://github.com/siddhesh/LuaJIT/tree/v2.1
The main LuaJIT repository is not seeing too many updates and the
maintainer has not been reviewing patches for some time now.  This
fork is primarily a means to consolidate abandoned but useful patches
to luajit that have been lying around in various trees such as
OpenResty, RaptorJIT (these will need extra work since it's a hard
fork) and architecture forks such as ppc64.

This tree also incorporates the LuaJIT-test-cleanup project into the
main repository to enable testing of the built binary.